backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
gunicorn==20.1.0
PyYAML==6.0
